+++ examples/evaluators/__init__.py
@@ -11,22 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-from .aime import AIME2024Evaluator, AIME2025Evaluator, AIMEEvaluator
-from .evaluator import AsyncEnvFactory, EvalSample, Evaluator
-from .metrics import MetricFn, compute_pass_at_k
-from .registry import get_benchmark, list_benchmarks, register
-
-__all__ = [
-    "AIME2024Evaluator",
-    "AIME2025Evaluator",
-    "AIMEEvaluator",
-    "AsyncEnvFactory",
-    "EvalSample",
-    "Evaluator",
-    "MetricFn",
-    "get_benchmark",
-    "list_benchmarks",
-    "compute_pass_at_k",
-    "register",
-]